requirements.txt
# --- Core ---
anthropic>=0.40.0
httpx>=0.27.0
python-dotenv>=1.0.0
mcp[cli]>=1.26.0

# --- Recolección de datos ---
pytrends>=4.9.2
feedparser>=6.0.12
pyyaml>=6.0.3

# --- CLI y scheduling ---
click>=8.3.1
apscheduler>=3.11.2

# --- Opcionales: instalar manualmente si se desean ---
# pip install ntscraper>=0.4.0         # Twitter/X sin API oficial
# pip install TikTokApi>=7.3.2         # TikTok (requiere: playwright install chromium)
